Fix IrminQueue.add and IrminQueue.peek
We need to create a new revision for each blob commited in the database, and we should tag revisions instead of blobs.

Now the interface looks very similar to Git, but you don't have a tree-like filesystem (eg. just revision/blob). That's not needed for dealing with simple queues for now on, so we'll see later.

All tests pass.