Give the CLI some love.
It is now possible to control irminsule stores using the "irmin" commands. Basic commands seems to work (including the pretty export to a Graphviz drawing). All the backends are supported (ie. in-memory, filestystem, remote) and you if you start in daemon mode, you can control control your irminsule instance using a JSON protocol. In summary: it starts to looks pretty usable.

Next, need to add the funny commands, such as snapshot/revert and import/export (and the watches, when they will be implemented).