Expose a new functor in Irmin to easily create Proxy
A "proxy" is something which share the same interface as a complete irminsule store, but where all the core calls (included the root calls to the irminsule store) are actually resolved by an other irminsule instance. This is pretty useful when you want to implement an efficent remote interface as the high-level API functions will be resolved on the server instead letting the client do lot of calls itselfs.