Rename IrminTree to IrminView
Now the view acts as regular RW store, but the reads are done lazily and write are done in memory.

I can see 2 scenarios where you want to use views, and the current API should/will be able to support both of them:

1. a view is a subtree in the database: you create the view, and you populate it on-demand (ie. every time you do a read). You can write on it, do some stuff, and when you are happy, you commit the whole tree back to the databse.

2. a view is a tree-like representation of an OCaml value. So you want to be able to convert an ocaml value to a tree, without having to worry about the underlying database. When you have done serializing your value, you can then choose in which path(s) in the database you want to serialize it.

This is WIP, there are still missing part (the merging back to the store bits, mainly).