WIP Ir_bc compiles
Also change `create: unit -> t Lwt.t` to `create: unit -> t`. The handler should always be immediate (that's not the case for the Git backend, but I'll fix) and fast, as it just an handler to the persistent store.

However, we cannot really get ride of the [t] everywhere, because sometime there is actually some state in the handler. The example is Ir_bc where we have the current tag/head in there, and function to modify it using side-effects.