Compute the least common ancestors only when really needed
In some cases (ie. 2-way merges) we don't care about computing the
lcas and these can be very costly.

Also, computing the lcas can itself cause some merging (to get only 1
common ancestor) which can return a conflict. In that case, expose the
conflict in the user-defined function and let the user deal with it. I
guess we could also return the list of lcas and let the user deals
with.