Fix the invalidation of cache in the view
A view can either be a key (we just created the view using View.of_path),
a database tree (we have added some stuff on the view, but we don't know
the hash of the contents yet), or both (meaning that we have already
read and write the view).

Modifying a view is done using side effects. `Keys` becomes `Both` on reads,
and stuff becomes `Node` (tree) on write. What was missing was a way to
transform a `Both` into `Node` again (meaning we need to serialize the
structure again on the next write) when a leaf is updated. This is now
done.

The two unit tests has been crafted to exhibit bad cache behavior, they
were failing before that patch, the pass after so that looks fine. Using
side-effects for efficient caching of views is a bit sloppy though, not
sure it was a good idea at the first place. But well, not it seems to
be ok for now on ...

Thanks @gregtatcam for reporting the issue.