Start adding the `compare_and_set` logics
`Irmin.update` is now supposed to be atomic, although only the in-memory backend
fully implements this currently in (as we can use normal locks).

Now need to fill the TODO with filesystem locking (woohoo) and make `Irmin.merge`
also atomic. Not sure all the high-level operations need to be atomic, but at
least these should be reviewed and their semantics should be specified somewhere.