Add Irmin.watch_key to watch a given path in an Irmin tree
The streaming are not optimized for the HTTP backend (we still received all
the head updates for the current branch and only the client do the filtering).
It's not hard to make the server do the filtering instead, but that means that
we should open a new stream for every watched sub-path... not sure it is worth
in practice.