Add HTTP client and server version checks
The client puts its version in the X-IrminVersion header of the request.
By default, the server will serve the request anyway if the header is not
set, as we still want default browser connections to work fine.

The server puts its version in the `versions` field of the JSON objects that
it returns. The client can deal with that and print the correct error message
instead of failing miserably in case of API mismatch.

Fix #167