Change the JSON stream API
The previous API was hard to parse (#266) and wrong (#269). This commit tries to
fix that.

Now the stream is an (almost valid) JSON array which looks like:

[ {"stream":"start"}, {"version":"..."}, {"result":..},..,{"result":..},{"stream":"end"}]

almost valid as the spec doesn't allow multiple fields with the same name; but
most of existing implementation don't really care.

This needs a special release of ezjsonm (0.4.2)