Move BC operations that ignore the branch to BC.Repo
These are: branches, remove_branch, heads, watch_branches, import,
export, and task_of_head.

Note that "heads" used to ensure that the current branch's head was
included in the returned set (which made a difference for anonymous
branches). This feature has been removed. Ideally, we should use OCaml's
GC to track which anonymous branches are still being used and return all
of them.