En route pour 1.0: Major API revision
For Irmin 1.0, the main design principles are:
- the backends has to provide the smallest possible API
- the user-facing Irmin API can be different from the backend one and should
  be nice/easy to use.

In 0.* versions, one of the goal was to make the backend API a subset of the
Irmin API. This was an interesting design approach but it caused more issues
that expected. Be more relax now, and allow the two APIs to evolve separately:
the backend API could be as minimal as possible -- the Irmin API could be
as expressive as possible.