Remove the `Empty subtree and change Tree.empty signature
There were two ways to create an empty subtree:
- `Empty and;
- `Node <empty-node>

This was a bit confusing, so remove the `Empty case. This forces `empty`
to become a function, as the hash of the empty node depends on the backend
used. Ideally we would cache the empty node in the Repo.t value, but it is
a bit hard to do currently.