Consolodate put/post
Some APIs make a distinction between PUT and POST operations, but I think that
this API is simple enough that such a distinction is artificial and will only
lead to confusion for the client. Therefore this commit merges create/update
operations into the same operation, which uses the POST command.