irmin-http: client and server should agree on how to serialise nodes and commit
This is needed to be able to cache things efficiently on the client.
As a side-effect, it makes Irmin_http.Make functor signature nice, e.g
now it's just  Irmin.S -> Irmin.S, which makes senses: the client just
proxy write/read operations to the server, the rest should be unchanged.