Add reverse traversal to iter
Replaced breadth first search with depth first search in iter in order
to support traversal of the closure graph in both direction: when `rev =
false` then a node (and the edges to its predecessors) is treated when
it is first visited otherwise a node is treated only after all of its
predecessors have been treated first. Skip and level are verified when
first visiting each node.