Do not fail when inode not found
Inodes cache [find t], where [t] is the node store on which a particular
inode is created. This means that if an inode points to a node from
another store, it will not find it. Do not fail in this case, to allow
some part of the tree to be stored somewhere else.