Test and fix representation uniqueness of inode
Fix a bug in `inode` where the `remove` function could cause hashing instabilities. No user-facing change since this function is not being used yet.

-- Size thresholds to trigger that bug
For directories containing at most `Conf.stable_hash` entries (by default 256), no bug occurred. In this case, this change does not induce any user-facing change.

For directories containing more than `Conf.stable_hash` entries, the directory's hash might depends on the past sequence of operations and is thus not deterministic, which breaks various Irmin invariants and could cause data inconsistencies.

For directories containing more than `Conf.entries` (32) entries but at most `Conf.stable_hash` (256) a few hashes internal to `inode` could have been inconsistent, but not the directory's hash. Adding more entries to such an inode would most likely bring it to the stage where even the root hash is affected.

-- Technical description
In an inode tree, whenever a `remove` operation emptied a leaf, the hash of all the ancestors of that leaf calculated a hash that was not the same as if that leaf never contained any values in the first place. The root hash is not affected on buggy inodes containing less than `Conf.stable_hash` entries because in that case the root hash is computed without relying on the structure if the tree, only the leaves are taken into account.