Use O_EXCL flag with O_CREAT
From the doc: “If O_CREAT and O_EXCL are set, *open*() shall fail if the
file exists. […] If O_EXCL is set and O_CREAT is not set, the result is
undefined.” We should then use the O_EXCL only after checking that the
file does not exists and when we want to create it.
The previous use of O_EXCL was wrong but it was not failing because it
wasn’t using the O_CREAT flag.