Revert "irmin-pack: try to call merges less often"
This reverts commit 587e219351e9a147cb7e3cea281ab025b61098c2.

The implementation of `Index.try_merge` is broken: it will attempt a
merge even when one is already running. As a result, whenever a batch
was completed, the main thread blocked on the completion of an ongoing
merge. This restores the old behaviour until a fix can be written and
tested in Index.

(cherry picked from commit 752b86583b89342901e05ec4cd6780a7342fe47c)