Avoid unnecessary Make indirections for multi-argument functors
This partially reverts changes introduced by
046ec54f38103e00561fa71144e6466e977aa63c where they are not strictly
necessary.  In particular, it avoids indirections of the following type:

```ocaml
module type Maker : sig
  module Make (...) = struct ... end
end
```

where they are not necessary (whenever the `Maker` type contains only a
`Make` functor and no other types).

This introduces some inconsistency in the `Maker` convention: when the
corresponding functor must perform a module substitution as part of
application then the old convention is still necessary, but overall it
reduces the boilerplate significantly in the common case.