irmin: avoid implicit coercions of dangling hashes to empty nodes
The existing implementation of `Tree` contains several operations that
consider "dangling" hashes (internal tree hashes that are `Not_found` in
the underlying store) to be equivalent to empty nodes. For instance:

- `Tree.find` at a path that contains a dangling hash will return
  `None` (and similarly `Tree.mem` returns `false`).

- `Tree.update` at a path that contains a dangling hash will observe a
  false empty node at the end of the path (... with similar results for
  `add` / `remove` etc.)

This behaviour can be triggered by the user using `Tree.shallow`, but
also arises when the underlying store is corrupt (and certain internal
objects can't be found). In each case, silently coercing to an empty
node is surprising, and so this commit changes these operations to raise
a `Dangling_hash` exception when this happens instead.