irmin: generalise contents keys to support inlining
This commit introduces a distinction between the type of contents keys
and node / commit key types. The former now takes a type parameter
corresponding to the type of the _value_ that the key points to,
allowing content keys to inline the value inside the key.

In future, we may want to add this capability to the other key types as
well, but this is more complex to implement as it introduces a cycle:
i.e. node values contain node keys, which may then contain recursive
node values.