irmin: add support for pruned tree objects
A "pruned" tree object is a purely in-memory tree with a particular
hash. Such trees can be used as children of other in-memory tree nodes,
for instance in order to compute the hash of the parent, but they cannot
be dereferenced.

Any operation that would require loading the contents of a pruned node
raises a new `Pruned_hash` exception. Attempting to export a tree
containing pruned sub-trees will fail similarly.