Use the explicit unboxed prehash
The index in inodes uses the unboxed step (by calling short_hash which
in turn calls pre_hash - the step is at the top level, and so it relied
on the pre_hash to unbox it). In the tests, we expect the contents to be
unboxed - the hash of the contents relies on the prehash of the string
primitive type.