irmin-pack: add support for user-defined indexing strategies
Now that internal references between pack entries can be resolved
without use of the index, it is no longer necessary to index every
object for correctness of the store. The index is now used to
de-duplicate exported (concrete) values, which impacts only the size of
the store and not its consistency.

This commit allows the user to define their own indexing strategy, with
two pre-provided strategies:

- `always`: equivalent to the old behaviour;
- `minimal`: index only the objects that are necessary for store
  correctness (currently just commit objects).

The `irmin-pack` test suite is now run twice, once for each of these
pre-provided strategies.