irmin-pack: use one-indexed versioning for pack value version
Renames the original object types as `*_v1`, and the new ones (with
added length headers) as `*_v2` to be consistent with the one-indexing
used for pack _store_ version headers. Now a V1 pack store is a file
containing only V1 objects, and V2 pack stores may also contain V2
objects.