irmin: fix order of stream proof extender segments
The segment list contained in a stream proof extender node is computed
in reverse order, and previously was not being properly reversed at the
end of this process.

This commit fixes the bug and introduces a `Reversed_list.t` type to
make it impossible to forget to reverse the list before use in future.