irmin-pack: use mmap for the mapping file
This commit changes the mapping file to use mmap on load. The code to
save the mapping has also changed, to take account of mmap format. In
hangzhou benchmarks, this substantially improves the time taken for the
main process to switch to the new prefix/suffix created by the GC
process. For example, the maximum switch time during the benchmark was
0.5s, for a 166MB mapping file (with gap_tolerance=0), with the load of
the mapping file itself being effectively instantaneous.

mmap files are not portable across architectures with different
endianness. Almost surely users are on little-endian machines. However,
to protect against a user copying a little-endian context to a
big-endian machine, the mapping file name has been changed to include
the endianness of the host machine.

We use binary search within the mmap array, rather than constructing an
explicit OCaml Map.t. This makes the load time effectively instant,
whilst also substantially reducing the memory usage (we don't have to
maintain an explicit tree in memory). Read-only processes which also use
the mmap should share the same memory, further reducing the overall
memory usage when running multiple processes.

For the mapping_file we add a private type to enforce invariants on
mmap'ed data (that the array has a size a multiple of 3, since the array
holds 3-tuples).

There is an additional test for [nearest_leq], which implements binary
search directly on an array to find the nearest key
less-than-or-equal-to a given key.

The existing [test_mapping] has been fixed to take account of the new
mapping-as-mmap implementation.

This replaces a previous PR #1966, and hopefully addresses most of the
comments from there.