irmin-pack: use mmap to load the mapping file
This commit changes the mapping file to use mmap on load. The code to
save the mapping has also changed, to take account of mmap format.

mmap files are not portable across architectures with different
endianness. Almost surely users are on little-endian machines. However,
to protect against a user copying a little-endian context to a
big-endian machine, the mapping file name has been changed to include
the endianness of the host machine.