Added tests that raise Invalid_depth exn/error
This exception is raised in irmin-pack when working
with intermediate or concrete inodes.

To surface this exception for intermediate inodes the
Internal.Raw.depth function has been mocked to return an
incorrect value.

For concrete inodes an incorrect tree is first constructed,
which, when loaded using `of_concrete`, causes the exception to
be raised which is caught and asserted as the equivalent
error type.